Updated TEST_Darkness scene, Player prefab
Updated player Sprites and material so correct sizing and lighting apply. Also changed the collisionbox size, as well added animator component to the player prefab.
Will need to look into separating animation dependency in the future, but this will allow use of the player for testing for the moment.
TEST_Darkness scene, messed around with the overworld light to see how dark a stage we might want.